+++ src/events/index.ts
@@ -1,0 +1,5 @@
+import * as EventEmitter from "./EventEmitter";
+
+const eventEmitter = new EventEmitter();
+
+export = eventEmitter;